+++ app/Interfaces/WordRepositoryInterface.php
@@ -7,7 +7,7 @@ use App\Models\Word;
 
 interface WordRepositoryInterface
 {
-    public function getAllWordsWithTranslations(int $perPage, ?string $cursor): CursorPaginator;
+    public function getAll(int $perPage, ?string $cursor): CursorPaginator;
     public function get(int $id): ?Word;
     public function create(array $data): Word;
     public function update(Word $word, string $englishWord): ?Word;