+++ app/Repositories/WordRepository.php
@@ -10,7 +10,7 @@ class WordRepository implements WordRepositoryInterface
 {
     public function __construct(protected Word $model){}
 
-    public function getAllWordsWithTranslations(int $perPage, ?string $cursor): CursorPaginator
+    public function getAll(int $perPage, ?string $cursor): CursorPaginator
     {
         return $this->model
           ->with('translations:id,word_id,spanish_word,german_word')
@@ -41,7 +41,6 @@ class WordRepository implements WordRepositoryInterface
 
     public function delete(Word $word): bool
     {
-        $word->translations()->delete();
         return $word->delete();
     }
 }